Refactor things to a single top-level Worker
