+++ workers/rust/container/src/lib.rs
@@ -1,7 +1,7 @@
 use bollard::Docker;
-use slog::{info, Logger};
+use slog::Logger;
 use taskcluster::Credentials;
-use taskcluster_lib_worker::claim::WorkClaimer;
+use taskcluster_lib_worker::Worker;
 
 mod executor;
 
@@ -11,12 +11,10 @@ pub async fn main(logger: Logger) -> anyhow::Result<()> {
     // TODO: load config
     let root_url = "https://dustin.taskcluster-dev.net";
 
-    info!(logger, "Starting Worker");
-
     let docker = Docker::connect_with_local_defaults()?;
     let executor = ContainerExecutor::new(docker);
 
-    let wc = WorkClaimer::new(executor)
+    let w = Worker::new(executor)
         .logger(logger)
         .root_url(root_url)
         .worker_creds(Credentials::from_env()?)
@@ -24,7 +22,6 @@ pub async fn main(logger: Logger) -> anyhow::Result<()> {
         .worker_group("rust".to_owned())
         .worker_id("worker".to_owned())
         .capacity(4)
-        .build();
-    let wc = wc.start();
-    wc.await
+        .start();
+    w.await
 }
